web/requirements.txt
flask
flask-mail
itsdangerous
pymongo
cerberus
pyjwt
typing-extensions
python-dotenv